Change carbon values to float to represent mmt (million metric tons)

+++ app/models/company.rb
@@ -6,8 +6,8 @@
 #  name               :string
 #  created_at         :datetime         not null
 #  updated_at         :datetime         not null
-#  total_co2_produced :integer          default(0)
-#  total_co2_captured :integer          default(0)
+#  total_co2_produced :float            default(0.0)
+#  total_co2_captured :float            default(0.0)
 #
 class Company < ApplicationRecord
   has_many :tasks
@@ -16,7 +16,7 @@ class Company < ApplicationRecord
 
   validates :name, uniqueness: true, presence: true
 
-  sig { params(year: Integer).returns(Integer) }
+  sig { params(year: Integer).returns(Float) }
   def co2_produced_in(year)
     self.company_years.where(year_id: year).first.co2_produced
   end
